+++ Banking-System/src/main/java/com/sandeep/mani/controller/CustomerController.java
@@ -1,4 +1,4 @@
-package com.sandeep.mani.customers;
+package com.sandeep.mani.controller;
 
 import java.util.List;
 
@@ -10,6 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.sandeep.mani.model.Customer;
+import com.sandeep.mani.service.CustomerService;
 
 @RestController
 public class CustomerController {
