--- Banking-System/src/main/java/com/sandeep/mani/employee/package-info.java
@@ -1,1 +1,0 @@
-package com.sandeep.mani.employee;
