--- Banking-System/src/main/java/com/sandeep/mani/identities/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author Sandeep Mani
- *
- */
-package com.sandeep.mani.identities;
