+++ Banking-System/src/main/java/com/sandeep/mani/model/Manager.java
@@ -1,0 +1,5 @@
+package com.sandeep.mani.model;
+
+public class Manager extends Employee {
+	
+}
