+++ Banking-System/src/main/java/com/sandeep/mani/model/Officer.java
@@ -1,0 +1,5 @@
+package com.sandeep.mani.model;
+
+public class Officer extends Employee {
+
+}
