+++ Banking-System/src/main/java/com/sandeep/mani/repository/CustomerRepository.java
@@ -1,6 +1,8 @@
-package com.sandeep.mani.customers;
+package com.sandeep.mani.repository;
 
 import org.springframework.data.repository.CrudRepository;
+
+import com.sandeep.mani.model.Customer;
 
 public interface CustomerRepository extends CrudRepository<Customer, String>{
