--- Banking-System/src/main/java/com/sandeep/mani/transaction/package-info.java
@@ -1,1 +1,0 @@
-package com.sandeep.mani.transaction;
